12/26/2025: Getting closer to resolving the jwt enabled checks. I'm almost there, but I'm probably going to have to get rid of my DevImpersonationFilter.

+++ src/main/java/com/monday/monday_backend/memory/entity/SessionOptionsEntity.java
@@ -26,7 +26,7 @@ public class SessionOptionsEntity {
 
     @Setter
     @Column(name = "max_chunks_per_session")
-    private Integer maxChunksPerSession;
+    private Long maxChunksPerSession;
 
     @Setter
     @OneToOne(mappedBy = "options", fetch = FetchType.LAZY)
