+++ src/main/resources/db/migration/V5__create_session_options.sql
@@ -4,5 +4,5 @@
 CREATE TABLE session_options (
   id                  UUID PRIMARY KEY DEFAULT gen_random_uuid(),
   session_scope       VARCHAR(255) NOT NULL,
-  max_chunks_per_session INTEGER
+  max_chunks_per_session BIGINT
 );
